+++ unittest/libmariadb/fetch.c
@@ -286,7 +286,7 @@ static int test_fetch_offset(MYSQL *mysql)
   stmt= mysql_stmt_init(mysql);
   FAIL_IF(!stmt, mysql_error(mysql));
 
-  rc= mysql_stmt_prepare(stmt, query, strlen(query));
+  rc= mysql_stmt_prepare(stmt, query, (unsigned long)strlen(query));
   check_stmt_rc(rc,stmt);
 
   memset(my_bind, '\0', sizeof(my_bind));
